+++ pkg/imgpkg/imageset/tar_image_set.go
@@ -136,6 +136,7 @@ func (i *TarImageSet) Import(path string, importRepo regname.Repository, registr
 		//crane.SaveOCI(*imgOrIndexes[0].Image, "/Users/ashishkumarsingh/Desktop/stuff/ashpect/imgpkg/cmd/imgpkg/a0")
 	} else {
 		imgOrIndexes, err = imagetar.NewTarReader(path).Read()
+		fmt.Println(imgOrIndexes[0].Ref())
 		//for testing
 		crane.SaveOCI(*imgOrIndexes[0].Image, "/Users/ashishkumarsingh/Desktop/stuff/ashpect/imgpkg/cmd/imgpkg/a0")
 		//crane.SaveOCI(*imgOrIndexes[1].Image, "/Users/ashishkumarsingh/Desktop/stuff/ashpect/imgpkg/cmd/imgpkg/a1")
