+++ pkg/imgpkg/imagetar/tar_reader.go
@@ -6,9 +6,11 @@ package imagetar
 import (
 	"fmt"
 	"io"
+	"os"
 
 	"github.com/google/go-containerregistry/pkg/name"
 	v1 "github.com/google/go-containerregistry/pkg/v1"
+	"github.com/google/go-containerregistry/pkg/v1/layout"
 	"github.com/vmware-tanzu/carvel-imgpkg/pkg/imgpkg/imagedesc"
 	"github.com/vmware-tanzu/carvel-imgpkg/pkg/imgpkg/imageutils/verify"
 )
@@ -147,4 +149,80 @@ func (r TarReader) getIdsFromManifest(file tarFile) (*imagedesc.ImageRefDescript
 		return nil, err
 	}
 	return ids, nil
+}
+
+func (r TarReader) ReadOci() ([]imagedesc.ImageOrIndex, error) {
+
+	//Check if the path is a OCI layout directory
+	stat, err := os.Stat(r.path)
+	if err != nil {
+		return nil, err
+	}
+
+	if !stat.IsDir() {
+		//give error "not a directory"
+		return nil, err
+	}
+
+	//TODO : FromPath checks for index.json but does not check for oci-layout, so add a check for oci-layout here.
+
+	//Get the oci layout rooted in the file system at path, layout index struct
+	l, err := layout.FromPath(r.path)
+	if err != nil {
+		return nil, err
+	}
+
+	ImageIndex, err := l.ImageIndex()
+
+	digest, err := ImageIndex.Digest()
+
+	//fmt.Println("ImageIndex's digest :", digest)
+
+	myImageIndex := imagedesc.ImageIndexIntermediate{
+		Index: ImageIndex,
+	}
+
+	fmt.Println("ImageIndex's tag :", myImageIndex.Tag())
+	fmt.Println("ImageIndex's ref :", myImageIndex.Ref())
+
+	//convert digest into a string
+	digestStr := digest.String()
+
+	ref := "index.docker.io/" + "ashpect/testrepo22@" + digestStr
+
+	//fmt.Println("Ref to be updated :", ref)
+
+	myImageIndex.SetRef(ref)
+
+	var i imagedesc.ImageIndexWithRef = myImageIndex
+
+	imageOrIndex := imagedesc.ImageOrIndex{
+		Image: nil,
+		Index: &i,
+		Labels: map[string]string{
+			"label1": "value1",
+			"label2": "value2",
+		},
+		OrigRef: "original-reference",
+	}
+
+	//add imageOrIndex to the slice of imageOrIndex
+	var imageOrIndexSlice []imagedesc.ImageOrIndex
+	imageOrIndexSlice = append(imageOrIndexSlice, imageOrIndex)
+
+	//imgOrIndexes = append(imgOrIndexes, imagedesc.ImageOrIndex{Index: ImageIndex})
+	//Handle multiples cases when manifests in index.json are >1
+
+	//IMP
+	//ok is the bool that tells us if the image is an image or an index
+	//t is the v1.Image or v1.ImageIndex
+	// t, _ := ImageIndex.(v1.Image)
+	// if err != nil {
+	// 	return nil, err
+	// }
+	// ----> file := tarFile{r.path}
+
+	//crane.SaveOCI(t, "/Users/ashishkumarsingh/Desktop/stuff/ashpect/imgpkg/cmd/imgpkg/hotstuff")
+
+	return imageOrIndexSlice, nil
 }
